aula_30_09_2022

+++ dsd_ms_worker_app/src/main/java/br/edu/ifms/dsd_ms_worker_app/controller/WorkerController.java
@@ -25,7 +25,7 @@ public class WorkerController {
     }
 
     @GetMapping(value = "/{id}/salary")
-    public ResponseEntity<Float> getWorkerSalary(@PathVariable("id") int id) {
+    public ResponseEntity<Double> getWorkerSalary(@PathVariable("id") int id) {
         Worker worker = this.workerService.getWorkerById(id);
         return ResponseEntity.ok(worker.getSalary());
     }
